02_ml_basics/wig20_xgboost_forecast/requirements_37.txt
pandas==1.3.5
numpy==1.21.6
scikit-learn==1.0.2
xgboost==1.5.2
pyyaml==6.0
matplotlib==3.5.3
pytest==7.0.1
fastapi==0.88.0
uvicorn==0.20.0
